refactor: merge High/Low EventConsumer, remove dead CubeRatePort (#902 MEDIUM)
- DUP-2: HighPriorityEventConsumer + LowPriorityEventConsumer (84 lines each)
  → IntegrationEventConsumer base class + thin @Component subclasses (14 lines each)
- DUP-9: Remove unused CubeRatePort (core.calculator.port) — zero references
- PAT-1/PAT-2: Skipped — ROI too low (~3 lines/file savings)
